Reorganize imports for PEP8 compliance
