+++ jan_simple_gui.py
@@ -1,7 +1,6 @@
 import tkinter as tk
-from tkinter import messagebox
+
 import requests
-import json
 
 class SimpleJanGUI:
     def __init__(self):
