+++ src/core/memory.py
@@ -2,13 +2,13 @@
 Memory management system for Jan Assistant Pro
 """
 
-import json
-import os
-import sqlite3
-import threading
 from contextlib import contextmanager
 from datetime import datetime
+import json
+import os
 from pathlib import Path
+import sqlite3
+import threading
 from typing import Any, Dict, List, Optional, Tuple
