+++ src/gui/__init__.py
@@ -1,3 +1,3 @@
-from .enhanced_widgets import ChatInput, EnhancedChatDisplay, StatusBar
+from src.gui.enhanced_widgets import ChatInput, EnhancedChatDisplay, StatusBar
 
 __all__ = ["StatusBar", "ChatInput", "EnhancedChatDisplay"]
