+++ src/tools/file_tools.py
@@ -3,8 +3,8 @@ File operation tools for Jan Assistant Pro
 """
 
 import os
-import shutil
 from pathlib import Path
+import shutil
 from typing import Any, Dict, List, Optional
