+++ tests/test_gui_widgets.py
@@ -1,12 +1,11 @@
 import os
 import sys
+import tkinter as tk
 import unittest
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src')))
 
-import tkinter as tk
-
-from gui.enhanced_widgets import ChatInput, StatusBar
+from src.gui.enhanced_widgets import ChatInput, StatusBar
 
 
 def create_root_or_skip():
